+++ app/ui/tray/browser.js
@@ -30,8 +30,6 @@ handlebars.registerHelper('i18n', function(key) {
 
 var sync;
 var syncStatus    = true;
-var showReset     = true;
-var showSync      = true;
 var showLogin     = true;
 var refreshQuota  = false;
 
@@ -82,25 +80,7 @@ function buildMenu() {
 
   menu.append(new MenuItem({type: 'separator', enabled: false}))
 
-  if(clientConfig.get('context') === 'development') {
-    if(showSync === true) {
-      label = i18n.__('tray.menu.startSync');
-      menu.append(new MenuItem({label: label, click:function(){
-        ipcRenderer.send('sync-start');
-        ipcRenderer.send('tray-hide');
-      }}))
-    }
-
-    if(showReset === true) {
-      label = i18n.__('tray.menu.resetSync');
-      menu.append(new MenuItem({label: label, click:function(){
-        showReset = false;
-        showSync = false;
-        ipcRenderer.send('dev-reset');
-        ipcRenderer.send('tray-hide');
-      }}))
-    }
-  } else if(clientConfig.get('loggedin') === true) {
+  if(clientConfig.get('loggedin') === true) {
     if(syncStatus === true) {
       label = i18n.__('tray.menu.pauseSync');
     } else {
@@ -174,22 +154,10 @@ ipcRenderer.on('link-account-result', (event, result) => {
 
 ipcRenderer.on('sync-started' , function() {
   syncStatus = true;
-  showSync = false;
-});
-
-ipcRenderer.on('sync-ended' , function() {
-  showSync = true;
 });
 
 ipcRenderer.on('sync-paused' , function() {
   syncStatus = false;
-});
-
-ipcRenderer.on('dev-reset-complete', (event, err) => {
-  if(err) return console.error('ERROR, reset not successfull', err);
-  showReset = true;
-  showSync = true;
-  return console.info('Reset complete', new Date());
 });
 
 ipcRenderer.send('tray-window-loaded');